Calculates and reports the emissions shadow price.
For each emission limit constraint in the model, this
commit calculates the corresponding shadow price and
reports it in a new database table.

Specifically, the commit directs the pyomo instance to
save the dual variables in memory (this is done in
temoa_run.py). Then, when formatting the results, the
model accesses the dual variable for each emission
limit constraint. These shadow variables correspond to
the corresponding implicit emission price.

License: GPLv2