Add Min and Max Seasonal Activity Constraints
Two constraints are added to the model:
- MinSeasonalActivity
- MaxSeasonalActivity
Both constraints are indexed over Regions, Periods,
Seasons, and Technologies.
The implementation is similar to MinActivity and
MaxActivity, only distinct values can be given per
representative day (i.e. 'season').
The motivation is to be able to prescribe daily
energy budgets for dispatch and energy limited
resources, such as hydro generators.

License: GPLv2