Add jobs reporting feature
This commit adds a feature that allows users to track
employment as a function of installed capacity. This
addition only affects model inputs and outputs -- i.e.
the optimization problem itself is not affected.

Inputs, in units of number of jobs per unit capacity,
are to be entered in the "EmploymentPerCapacity" table
in the SQLite file.

Outputs are printed to the "Output_Employment" table in
the SQLite file.

The output, in units of number of jobs, is the product
of the jobs per unit capacity and the installed capacity.