Remove symbolic variable names when using CBC
This commit removes the ability to use symbolic solver
labels when using the CBC solver. This is because the CBC
solver produces an error when variable names exceed
100 characters, which is not checked for in Temoa.

The result is that the generated LP files now use generic
variable names, e.g. x1200052, in place of symbolic names,
e.g. c_a_DemandConstraint_Demand1_2020_04_H01.