Add variable intra-annual efficiency feature
This commit adds a feature to vary technology efficiencies at the
time step level. Prior to this commit, technology efficiencies were
indexed over their vintage only -- i.e. it was not possible to vary
an efficiency throughout a year.

This is accomplished by adding an additional table to the SQL/SQLite
files: EfficiencyVariable. The EfficiencyVariable parameter is indexed
over {region, input commodity, technology, season, time of day, and
output commodity} whereas the Efficiency parameter is indexed over
{region, input commodity, technology, vintage, and output commodity}.

Notice the EfficiencyVariable parameter is not indexed over vintage.
This means that all vintages of the technology in question are assumed
to have the same intra-annual variation in efficiency.

The default value for the EfficiencyVariable parameter is 1.

The implementation works as follows: If you would like to have technology
A's efficiency vary throughout the year, begin by inputting its nominal
(or annual average) efficiency in the Efficiency table as before. Then,
use the EfficiencyVariable table to input factors for each time step of
the year that, when multiplied by the technology's nominal efficiency,
yield the expected efficiency.

For instance, say the technology in question is a heat pump with a seasonal
coefficient of performance (i.e. efficiency) of 3.0, but you would like for
this COP to be 2.0 on cold days and 3.3 on mild days. You would accomplish this
by setting its efficiency to 3.0 in the Efficiency table, and by setting the
EfficiencyVariable value to 0.667 for all "cold day" time steps and 1.1 for
all "mild day" timesteps.

Important: this feature does not apply to technologies in the "tech_annual"
set -- i.e. technologies with constant annual output (see the Temoa documentation
for more information). These technologies assume their nominal efficiency in all
time steps.