Allow Duals reporting to be optional (copy)
Note: This commit is a copy of the commit TemoaProject/temoa@efc73ff
on the original TemoaProject/temoa repository from which
this repository is based. The original commit was made by
Aranya Venkatesh and its message is pasted below:

"Through settings in the config file, the user can specify
if Duals are to be reported to the .sqlite file. The default
settings are False, i.e., they are not reported."