Fix bug in myopic solve that deletes extra techs (copy) Note: This commit is a copy of the commit TemoaProject/temoa@1831e3a30f780f721e4898d27457650e95ce70a7 from the original TemoaProject/temoa repository from which this repository is based. The original commit was made by Aranya Venkatesh and its message is pasted below:
"During the myopic solve, processes that retire before the start of the time
period currently being solved are deleted. The logic previously coded did not
account for distinct region and tech names which resulted in a few processes
being inadvertently deleted. This was updated with a new sql query in
temoa_model/temoa_myopic.py"