Remove small values of primary/linked techs from myopic solution (copy)
Note: This commit is a copy of the commit TemoaProject/temoa@c90e029d02e37b18d2ced97c89361877cd7c484d
on the original TemoaProject/temoa repository from which
this repository is based. The original commit was made by
Aranya Venkatesh and its message is pasted below:

"The solvers sometimes report small values for primary technologies
but not the linked technology, or vice versa. This causes issues
in future period solves where the capacity for one of the technologies
exists but not the other. This code looks for such instances and removes
related data from the output tables."