Allow 'global' in min/max activity constraint (copy)
Note: This commit is a copy of the commit TemoaProject/temoa@e400b4e3b6b527f0ee36aa0775cf72235adb737c
on the original TemoaProject/temoa repository from which
this repository is based. The original commit was made by
Aranya Venkatesh and its message is pasted below:

"Allow MinActivity and MaxActivity tables to include rows with 'global'
as a region, where the activity variables are summed over all regions
in the database. These activity variables do not include those relating
to regional exchange processes.
temoa_myopic.py is modified as well, to make sure that rows containing
'global' as a region are not deleted, since they don't specfically appear
in the Efficiency table."