Change region separator from "-" to ","
Previously, the region separator token for constraints and parameters
was the "-" character. That is, to set a constraint for a set of
regions, you would enter "NY-ME-CA".

This has been changed to a ",". This is because Temoa creates
transmission regions, which are treated as single regions, by
using a "-" between the two regions with the transmission line.

This means "NY-NJ" should be parsed as a single region, not two
distinct regions.