Add Min/MaxNewCapacityShare
These constraints allow the user to set a minimum or maximum
share of new capacity for a given technology within a technology
group to which it belongs.

For instance, you might define a tech_group of light-duty vehicles, whose
members are different types for LDVs. This constraint could be used to enforce
that no less than 10% of new LDV purchases in a given year must be of a certain type.